refactor(resourceclient): move ObjectIdentityKeyVals to utils package
ObjectIdentityKeyVals was defined in the resourceclient package but used
across unrelated packages (finalizer, status, capprevision). Moving it
to the shared utils package removes the cross-package coupling and gives
it a more appropriate home alongside other general-purpose helpers.
