+++ internal/kinds/capp/resourceclient/resourceclient.go
@@ -6,6 +6,8 @@ import (
 
 	"github.com/go-logr/logr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	utils "github.com/dana-team/container-app-operator/internal/kinds/capp/utils"
 )
 
 type ResourceManagerClient struct {
@@ -13,26 +15,9 @@ type ResourceManagerClient struct {
 	Log       logr.Logger
 }
 
-// ObjectIdentityKeyVals returns key/value pairs identifying obj for structured logs.
-func ObjectIdentityKeyVals(obj client.Object) []any {
-	gvk := obj.GetObjectKind().GroupVersionKind()
-	kind := gvk.Kind
-	if kind == "" {
-		kind = "Unknown"
-	}
-	return []any{
-		"kind", kind,
-		"group", gvk.Group,
-		"version", gvk.Version,
-		"namespace", obj.GetNamespace(),
-		"name", obj.GetName(),
-		"resourceVersion", obj.GetResourceVersion(),
-	}
-}
-
 // CreateResource creates a resource.
 func (r ResourceManagerClient) CreateResource(ctx context.Context, resource client.Object) error {
-	r.Log.Info("kubernetes API write create", ObjectIdentityKeyVals(resource)...)
+	r.Log.Info("kubernetes API write create", utils.ObjectIdentityKeyVals(resource)...)
 	if err := r.K8sClient.Create(ctx, resource); err != nil {
 		return fmt.Errorf("failed to create resource %s %s: %w", resource.GetObjectKind().GroupVersionKind().Kind, resource.GetName(), err)
 	}
@@ -41,7 +26,7 @@ func (r ResourceManagerClient) CreateResource(ctx context.Context, resource clie
 
 // UpdateResource updates a resource.
 func (r ResourceManagerClient) UpdateResource(ctx context.Context, resource client.Object) error {
-	r.Log.Info("kubernetes API write update", ObjectIdentityKeyVals(resource)...)
+	r.Log.Info("kubernetes API write update", utils.ObjectIdentityKeyVals(resource)...)
 	if err := r.K8sClient.Update(ctx, resource); err != nil {
 		return fmt.Errorf("failed to update %s %s: %w", resource.GetObjectKind().GroupVersionKind().Kind, resource.GetName(), err)
 	}
@@ -50,7 +35,7 @@ func (r ResourceManagerClient) UpdateResource(ctx context.Context, resource clie
 
 // DeleteResource deletes a resource.
 func (r ResourceManagerClient) DeleteResource(ctx context.Context, resource client.Object) error {
-	r.Log.Info("kubernetes API write delete", ObjectIdentityKeyVals(resource)...)
+	r.Log.Info("kubernetes API write delete", utils.ObjectIdentityKeyVals(resource)...)
 	if err := r.K8sClient.Delete(ctx, resource); err != nil {
 		return fmt.Errorf("failed to delete %s %s: %w", resource.GetObjectKind().GroupVersionKind().Kind, resource.GetName(), err)
 	}
