+++ internal/kinds/capprevision/adapters/capprevision_adapter.go
@@ -6,7 +6,7 @@ import (
 	"maps"
 
 	cappv1alpha1 "github.com/dana-team/container-app-operator/api/v1alpha1"
-	rclient "github.com/dana-team/container-app-operator/internal/kinds/capp/resourceclient"
+	utils "github.com/dana-team/container-app-operator/internal/kinds/capp/utils"
 	"github.com/go-logr/logr"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -69,9 +69,9 @@ func CreateCappRevision(ctx context.Context, k8sClient client.Client, logger log
 		return err
 	}
 
-	logger.Info("kubernetes API write create", rclient.ObjectIdentityKeyVals(&cappRevision)...)
+	logger.Info("kubernetes API write create", utils.ObjectIdentityKeyVals(&cappRevision)...)
 	if err := k8sClient.Create(ctx, &cappRevision); err != nil {
-		logger.Error(err, "failed to create CappRevision", rclient.ObjectIdentityKeyVals(&cappRevision)...)
+		logger.Error(err, "failed to create CappRevision", utils.ObjectIdentityKeyVals(&cappRevision)...)
 		return err
 	}
 	return nil
@@ -79,9 +79,9 @@ func CreateCappRevision(ctx context.Context, k8sClient client.Client, logger log
 
 // DeleteCappRevision deletes a specified CappRevision and returning an error on failure.
 func DeleteCappRevision(ctx context.Context, k8sClient client.Client, logger logr.Logger, cappRevision *cappv1alpha1.CappRevision) error {
-	logger.Info("kubernetes API write delete", rclient.ObjectIdentityKeyVals(cappRevision)...)
+	logger.Info("kubernetes API write delete", utils.ObjectIdentityKeyVals(cappRevision)...)
 	if err := k8sClient.Delete(ctx, cappRevision); err != nil {
-		logger.Error(err, "failed to delete CappRevision", rclient.ObjectIdentityKeyVals(cappRevision)...)
+		logger.Error(err, "failed to delete CappRevision", utils.ObjectIdentityKeyVals(cappRevision)...)
 		return err
 	}
 	return nil
